Add MainActivity with buttons, blank WeeklyReport
Add contents of MainActivity from another android studio project. The
MainActivity displays a progress bar and a clickable water bottle that
works as a button updating the progress bar and the text above the
progress bar. A button with the text "WEEKLY REPORT" transitions to the
WeeklyReport activity. The activity WeeklyReport was holds a TextView
that displays "This is the weekly report screen".

The updates to MainActivity accomplishes the functionality of a
clickable water bottle image that the user can press to update a
progress bar that shows the user's progress towards their daily goal.
The WeeklyReport activity will hold data about user activity and is
one of the main screens used in the app.