Add circle drawable
Defined a circle drawable using android:shape="oval" and setting the
width and height of the circle to the same value.

Circle are required for the functionality of highlighting circles
next to days of the week to show the user the days in the week that they
accomplished their intake goal.