Add onClick method for learn_more_button
Add the onClick method learnMoreOnClick to the learn_more_button. The
long toast message "Directs user to a web-page to learn more about the
benefits of drinking water" when the button is clicked.

Each button that is not implemented needs to display a toast message
to show what the button should do.