Add DatabaseHelper class to store progress
The DatabaseHelper class inherits from SQLiteOpenHelper and creates
a database called water_database that contains the tables goal_progress
and daily_input. The table goal_progress stores the date, goal, and
progress towards the goal. The table daily_input stores the date, time,
and amount that is entered.

Storing data in a SQLite database is used to know what days of the week
the water goal is met and records when water is tracked.