Add insertData and call insertData in waterOnClick
The insertData() method was added to the DatabaseHelper class and
inserts the data that is passed into the method into both of the tables
of the database. This method is called in the waterOnClick method to
update the data in the database when the user adds progress towards
the water goal.

The insertData() method is necessary to easily insert data into the
database when the user tracks water. The data recorded when the user
tracks water is used to update the progress shown on the weekly report
screen.