Populate data fields in WeeklyReport
Three listviews are used in WeeklyReport to get records of when the user
is tracking more water. The data in the listviews are populated using
the getDate(), getTime(), and getAmount() methods in the DatabaseHelper
class.

The listviews are used to show the user when they have recorded that
they have drank water.