Add web service to populate recommended goal text
Add GET request in android code through the GetWebServiceData class
to retrieve a value for recommended goal. The web service that is
queried is accessed through the url http://70.32.24.170:8080/goal
and the parameter weight can be passed into the web service to get
a different goal. The web service returns a goal value that is half
the weight value that is passed as a parameter. The recommended goal
is displayed through a text view that is displayed at the top of the
main activity.

Using a web service is a requirement for the project and adding a class
that accesses the web service and extends AsyncTask allows for data to
be retrieved.