Add methods to color code day of the week circles
The method getCompleteDays() was added to the DatabaseHelper class
and returns an integer array that contains the values for the days of
the week where the water goal for that day was met. In the MainActivity
the method setDayOfWeekColors takes in the array from getCompleteDays()
and changes the colors of the corresponding drawables to reflect which
day of the week the water goal was met.

Adding these methods provides information to the user about the days
that they are meeting their water intake goal.